[Priroda] refactor: move entry_fn from Miri bin to eval module to allow external access for Priroda use
